:bug: Persistable 프로토콜을 모든 Repository에서 채택할 수 있도록 변경
