+++ iOS/MSCoreKit/Sources/MSPersistentStorage/Protocol/MSPersistentStorage.swift
@@ -8,6 +8,7 @@
 public protocol MSPersistentStorage {
     
     func get<T: Codable>(_ type: T.Type, forKey key: String) -> T?
+    @discardableResult
     func set<T: Codable>(value: T, forKey key: String) -> T?
 
 }
